cast gyro angle to int

+++ src/Team167Lib/SwerveDrive.cpp
@@ -189,5 +189,5 @@ int SwerveDrive::GetWheelAngle()
 
 int SwerveDrive::GetGyroAngle()
 {
-	return gyro->GetAngle() % 360;
+	return (int) (gyro->GetAngle()) % 360;
 }
